[Feat] RestaurantSummariesMapper 생성 (#32)

+++ core/model/src/main/java/com/posomo/saltit/model/dto/RestaurantSummaryDto.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class RestaurantSummaryDto(
     @SerializedName("titleImageUrl") val titleImageUrl: String,
     @SerializedName("restaurantName") val restaurantName: String,
-    @SerializedName("rating") val rating: Int,
+    @SerializedName("rating") val rating: Float,
     @SerializedName("mainMenuPrice") val mainMenuPrice: Int,
     @SerializedName("mainMenuName") val mainMenuName: String,
     @SerializedName("categoryName") val categoryName: String,
